+++ plumbum/local_machine.py
@@ -664,6 +664,8 @@ class LocalMachine(object):
         def list_processes(self):
             """
             Returns information about all running processes (on Windows: using ``tasklist``)
+
+            .. versionadded:: 1.3
             """
             import csv
             tasklist = local["tasklist"]
@@ -681,6 +683,8 @@ class LocalMachine(object):
         def list_processes(self):
             """
             Returns information about all running processes (on POSIX systems: using ``ps``)
+
+            .. versionadded:: 1.3
             """
             ps = self["ps"]
             lines = ps("-e", "-o", "pid,uid,stat,args").splitlines()